+++ src/Data/CERES/Value.hs
@@ -11,6 +11,7 @@ data ValueTyper = ValueTyper
   , valueType      :: ValueType
   } deriving (Show, Eq)
 
+-- TODO: Can't determine whether `(ErrValue _) /= (ErrValue _)` or not
 data Value
   = IntValue { iV :: Int }
   | DblValue { dV :: Double }
@@ -26,6 +27,23 @@ instance Show Value where
   show (BoolValue bV) = "BV<| " ++ show bV ++ " |>"
   show (ErrValue bV) = "EV<| " ++ T.unpack bV ++ " |>"
 
+errValueWith2 :: (Show a, Show b) => Name -> Message -> a -> b -> Value
+errValueWith2 funcName errorType vA vB = ErrValue errorMessage
+  where
+    errorMessage = T.concat
+      [ "[Error]<"
+      , funcName
+      , " :=: "
+      , errorType
+      , "> "
+      , T.pack . show $ vA
+      , " and "
+      , T.pack . show $ vB
+      ]
+
+errValueTEWith2 :: (Show a, Show b) => Name -> a -> b -> Value
+errValueTEWith2 funcName = errValueWith2 funcName "TypeError"
+
 data ValueType
   = VTInt
   | VTDbl
@@ -40,3 +58,20 @@ instance Show ValueType where
   show VTStr  = "C-Str"
   show VTBool = "CBool"
   show VTErr  = "C-Err"
+
+-- Value Container for abstract vaule's real place and type
+data ValueContainer vp = VC
+  { valueID :: ID
+  , valuePlace :: vp -- ValuePlace
+  } deriving (Eq, Ord)
+
+instance Show vp => Show (ValueContainer vp) where
+  show vc = "<[" ++ show (valueID vc) ++ "@" ++ show (valuePlace vc) ++ "]>"
+
+data StandardValuePlace
+  = AtLocal
+  | AtTime
+  | AtWorld
+  | AtDict
+  | AtVar
+  deriving (Eq, Ord, Enum,  Show)