+++ ChangeLog.md
@@ -8,6 +8,14 @@
 
 Change type of `CERES` and `Value` as parametric to allow to change type of Str, etc.
 
+## [0.1.0.0] - 2019-05-06
+### Changed
+* Move functional modules from `Data.CERES.*` to `CERES.*`
+* Generalized instructions by `ValueContainer`
+* Rename `CERES.Operator` to `CERES.Operate`
+### Added
+* Implement `ValueContainer`
+
 ## [0.0.0.1] - 2019-02-23
 ### Added
 * Implement Data.Type